Entry points

Contour.contour — Function.

contour(x, y, z, level::Number) Trace a single contour level, indicated by the argument level.

You'll usually call lines on the output of contour, and then iterate over the result.

source

Contour.contours — Function.

contours returns a set of isolines.

You'll usually call levels on the output of contours.

source

contours(x,y,z,levels) Trace the contour levels indicated by the levels argument.

source

contours(x,y,z,Nlevels::Integer) Trace Nlevels contour levels at heights chosen by the library (using the contourlevels function).

source

contours(x,y,z) Trace 10 automatically chosen contour levels.

source

Accessors

Contour.level — Function.

level(c) Indicates the z-value at which the contour level c was traced.

source

Contour.levels — Function.

Turns the output of contours into an iterable with each of the traced contour levels. Each of the objects support level and coordinates.

source

Contour.lines — Function.

lines(c) Extracts an iterable collection of isolines from a contour level. Use coordinates to get the coordinates of a line.

source

Contour.coordinates — Function.

coordinates(c) Returns the coordinates of the vertices of the contour line as a tuple of lists.

source

Utilities

Contour.contourlevels — Function.

contourlevels(z,n) Examines the values of z and chooses n evenly spaced levels to trace.

source