Entry points

Contour.contour — Function

contour(x, y, z, level::Number)
contour(x, y, z, level::Number, VT::Type)

Trace a single contour level, indicated by the argument level. The extracted vertex type maybe be specified by VT.

You'll usually call lines on the output of contour, and then iterate over the result.

Contour.contours — Function

contours returns a set of isolines.

You'll usually call levels on the output of contours.

Accessors

Contour.level — Function

level(c) Indicates the z-value at which the contour level c was traced.

Contour.levels — Function

Turns the output of contours into an iterable with each of the traced contour levels. Each of the objects support level and coordinates.

Contour.lines — Function

lines(c) Extracts an iterable collection of isolines from a contour level. Use coordinates to get the coordinates of a line.

Contour.coordinates — Function

coordinates(c) Returns the coordinates of the vertices of the contour line as a tuple of lists.

Utilities

Contour.contourlevels — Function

contourlevels(z,n) Examines the values of z and chooses n evenly spaced levels to trace.